test-requirements.txt
flake8
django-stubs
mock
isort
typing_extensions
pytest
pytest-django
pytest-cov
